Dont' use K8S Model objects, but dictionaries
Using the official Kubernetes Model classes
(https://github.com/kubernetes-client/python) is avoided because it
presents some problems:

1) Custom Resource Definitions (eg.: kubeflow)
      => This is a BLOCKER for why it is not a good idea to use the
python-client as the provider for the model classes.

2) When using the to_dict() of the classes, that is not a valid
Kubernetes yaml since it outputs the fields as snake_case and not
camelCase as expected by the Kubernetes API.
      => This can be fixed with a workaround, so it is just an annoying
fact.

3) Hacky way to create the class from a YAML object.
       => There is no clean way to create a Kubernetes model object from
a yaml file without going through the internals of the Python client
API.
       The closest method is `create_from_yaml()`, which is analogous to
`kubectl apply -f file.yaml`, which:
       - calls the K8S Api server with the contents of the yaml.
       - returns the results as a Kubernetes Model object by
deserializing the response.

       Diving into the `create_from_yaml()` method I found this
`deserialize()` method, which can create the objects correctly. However
it requires a JSON as input and also the type it must create:

           def deserialize(self, response, response_type) ->
response_type:
              return self.__deserialize(json.load(response.data),
response_type)

       The theory is that these Kubernetes Model objects were not made
for being manipulated as real model objects, but only for having a
"type-safe" way of interacting with the responses of the Kubernetes
server.