Put the public subnets at the end of the CIDR spectrum
The intention is to later have as many private subnets as availability
zones so that EKS workers can be launched in all AZs and we can make use
of all available hardware in the region.

The problem is that Terraform doesn't handle conflicting ip ranges
between private/public subnets when applying the plan. The error that
Terraform shows when a subnet that was "private" and then becomes
"public" is:

```
InvalidSubnet.Conflict: The CIDR 'XXX' conflicts with another subnet
```

The reason is that inside the `vpc` module each "private"/"public"
subnet is handled by different `resource` blocks, and Terraform is not
smart enough to handle these cases.