Let Makefile template deployment more fully
In order to not have redundant bits of information everywhere that
make it painful for coordinating the different files that are
necessary to provide a build environment and make deployments, the
makefile is the central source of truth.  The top portion of the file
contains all the variables necessary to pin down the moving parts of
this application.  Along with Make, Sed is your friend.

Philosophically, I believe you should use simple and effective tools
that do one thing well and then it is up to you to coordinate and
orchestrate how they address your problem.