Workaround Terraform problem when using a remote state && Use a predictable name for the s3 bucket
# Change 1: Workaround Terraform problem when using a remote state

Terraform assumes it is running from within the root directory of the
modules (where `main.tf` is located) on many of its commands.

This problem does not show up when using a local state as we were doing
before, however, when using remote state (S3) then the problem shows up.

The bug has been reported multiple times to Terraform:
- https://github.com/hashicorp/terraform/issues/18030
- https://github.com/hashicorp/terraform/issues/16723

To which their answer is:

> running terraform against a directory that isn't the current directory
> is not fully supported

# Change 2: Make the name of the s3 bucket to be predictable

It is better to have a predictable s3 bucket name because it allows
"importing" the state on a new machine.

This required moving the `region` parameter handling into the baictl,
out of terraform because the region name is in the bucket name. This is
not strictly required, but it will allow us to create multiple clusters
in the same account later on. It also keeps the bucket within the same
region as specified by the user.