Make the name of the s3 bucket to be predictable
It is better to have a predictable s3 bucket name because it allows
"importing" the state on a new machine.

This required moving the `region` parameter handling into the baictl,
out of terraform because the region name is in the bucket name. This is
not strictly required, but it will allow us to create multiple clusters
in the same account later on. It also keeps the bucket within the same
region as specified by the user.