Make the constructor of the record take the channels
This is okay, but is a bit strange to me the way I am crossing over
functional and OO programming.  This whole thing feels a bit dirty to
me and begs many questions that I should not even have to ask because
of the situation.

Because these are fields in the record... I have the raw channels
available in this record (map) object that can be gotten to... but
only if you have the record map directly.  I don't want to go through
the process of passing it around to all of the places it is needed.
So I crate atoms of the channels in the namespace open for folks to
refer them into their namespace and use them.  This dichotomy is odd.

I find that I am shoe-horning complexity because of the OO shape of
this code, whereas it would have been a lot more straightforward as
straight-up functional code with raw open datastructures being acted
on by a bunch of functions.

I do like using the channel abstraction so that I can generate
messages from anywhere I can get a hold of the channel and not care
about the other end.  I went this path because my teammates dig OO
stuff, so it should be more readable to them.