Sewed the HTTP request to the Kafka dispatching
Using core.async channels the HTTP handler will dispatch
augmented/transformed input events from the clients to events sent to
Kafka on the BAI_APP_BFF channel.

Some things to note:
The transformation work is done in the events namespace however, the
calls for the hand off is done directly in the handler.  I do this
because it is the right place to catch any problems and be able to
provide immediate synchronous response to the client.  If there was
something wrong and we put that code in the kafka sink, we would not
have visibility.  Also the sinks and sources should have a clear
separation of concerns and not conflate event reception or sending
with transformation.  Following the general "ports and adapters"
design, the transformation would be done at the yet to be formally
created adapters layer.

https://herbertograca.com/2017/09/14/ports-adapters-architecture/

As this projects evolve so will this design to more formally represent
these notions.