Clean saffolding pre GraalVM
The main skeleton for the command-line parsing and delegation is in
place.  But now, before I go further I need to make sure that where we
are agrees with GraalVM.  I suspect that there be dragons when I start
to sew in the http stuff, so let's get GraalVM going next.