Speed up of dev container but there is a catch...
We want to speed up the dev container but putting in it all the things
that it would normally download.  This way they are fetched once and
cached with the magic of docker.  I noticed that the dev container was
consistently fetching this 14M leiningen-2.9.1-standalone.jar file.

So I put the jar in etc and uncommented the COPY line for it and
created an image.  The caveat comes from the fact that we don't want
to store this 14M artifact in our source tree, yet because of Docker
we can't reach beyond the root context to pull it in, of course.
The result is... you have to do the file move into etc manually and
uncomment the lines to build a more efficient image.  Otherwise, it
will be fine, you just have to pay the cost of downloading the lein
jar.