Refactored the status output fns + include initial submission status
Now there is an initial status that reflects what you are sending and
where you are sending it to.  This is so the user has a clear
understanding of the model, i.e. that they are sending their toml file
to an endpoint... and that they can quickly inspect that it is the
correct endpoint.