[bff] [WIP] record only successful submits to fs database
Only when the output from the submission is a UUID-ish value should we
then write down the value to our filesystem ("database") as well as
the status events.  Keep the status events in a hierarchy that
resembles that structure of the JSON stored in the BFF...
<client_id>
     <action_id>
         data (where the json events are recorded)

In this way we can efficiently distribute the space and make querying
much faster.