[bff] [WIP] Timestamps - record them, query for time since
In an effort to make it so that we don't have to go to the server
unless we have to, I am making it so that we will query since the
last point in time when we wrote an event down.

To follow up, must update the server to use query parameter "since".
Must also write events to file and THEN jq that file, instead of
Tee'ing the output. Of course will have to put in a --force mode that
makes it go to the bff for info.  Also have the ability to clear out
events and refresh.