[bff] [WIP] Status caching and more efficient data xfers
Basically we query local stores of data and only asks for the new data
from the server.  This means that the server won't send more data than
is necessary and the user can perform local operations on the cached
data.  Makes things more efficient and allows for system to scale that
much better.  It also makes for a great user experience because
everything is snappy.

This is not completed work.  I left a couple #TODOs there with next
steps. Not including the change in the service to accept a new query
parameter, "since".