[bff] Fetching status events "since" a given time
There is a bit of defensive coding, nil patching the call to
get-all-client-jobs-for-action, but pretty much the main thing is
supporting getting limited data back from the server, only since the
time specified by the HTTP API "since" query field.

Of course this atom does not suffice for the data structure because it
is in memory only and not distributed so it must be replaced by
something that is... not. :-)