[bff] Show TAG when viewing rollout history listing
We want to make sure that we keep the data provenance super clear
throughout the project.  We should always know what is running, down
to the commit in source code control.  This template file reflects the
tag as a label.  With this PR it also reflects the TAG during rollout
history lisiting

See:
%> make rollout-history