[bff] Addressed erroneous TODOs, but...
This is still not the right solution.  The right solution would be
describe defrecords (objects) that contained these two fields that get
filled in if set or use these default values.

So a TODO here is to define a "topic" is such a way that we can
individually set these values per topic.

@jlcontreras
:+1: