Expose BFF (#519)
* [bff] Service object in place (still needs tuning)

* Be consistent with the set of labels and selectors

Make sure that the set of labels I am selecting over in the service
matches the labels I am posting in the deployment so that the service
selector can match on the deployment.  I am including environment here
for safety (though we don't need it yet).  When environment becomes an
issue we want to safeguard that we don't nuke the wrong things!

(either remove enviroment everywhere or keep it everywhere)

* Using annotation for security group

* Put security group for service LB in an array

* updated to the "right" security group

* Using kubectl and sed in makefile to gen service.yml from template = FRESH

* IDEA rules one more time (#540)