[bff] More reliable Conda detection, Includes curl...
There was some issue with the system detecting Conda. Now we will
hinge this detection on the presence of CONDA_EXE environment
variable.

Besides some cosmetic changes we are actively asserting the version of
curl that we know works well with this script.

... misc cosmetic changes.

Note:
From previous commit we exect BASH to be at version 5+ or we don't
continue.  Also we no longer install bash through Conda because you
can't initiate a new bash in the bootstrap bash. (at least not that I
know of that works).  Having ONE dependency... is not asking much.