[bff] Anubis client is done.... mostly... ish(?)
There are still some "TODOs" to... erhm, *do*, however, I believe we
have the desired functionality to delight our users, at least in the
context of an initial release.

Okay... it is BASH... Don't hate.  Initially, as is usually the case,
it was a simple client to make some curl calls against a service... It
grew to being a bit more than that to address giving the users a
delightful experience.

Loose ends...
 * Rendering the command return events :shrug:
 * Still needs to call into the back end to get the raw results of the
 model run
 * And yes, the TODOs...