[bff] FUCK! S3 writing still does NOT work... disabling...
After working on permission stuff with Edison... still was not able to
successfully deploy the S3 writing feature.

I won't let this hold up the Anubis client work that is on top of this
so I am going to comment this out and get to it at a later date.  I've
burned too many minutes I'll never get back on this atm.  Deeply
frustrated.  And it is probably something silly with how i am making
the call or something unrelated.  Anyway... let's move on.

Error I get in the container:

```
19-07-09 16:35:28 bai-bff-6fdff56766-hlflx ERROR [bai-bff.http-api:45] - Could Not Parse Descriptor Input: Access Denied (Service: Amazon S3; Status Code: 403; Error Code: AccessDenied; Request ID: 6C4CBA2BB86B1AEB; S3 Extended Request ID: 4jvYlBmsFNt+vMIw24MdGrdStCJhnsTWgx9HKGhKC4XGZaXiQYldlhgtkpScnwc5Rl0iCJJbo/c=)
```

See commit: 68b9e5f and db5362d for the shenanigans.