Bff s3 dir vs uri (#622)
* [bff] S3 - insert a "folder" level

* [bff] Changed the way Anubis client constructs curl call

* A bit more cosmetic clarity to user feedback upon submission

* [bff] Create a way to force the script to be posted

* [bff] Clean up after yourself in containers - no state (sort of)

* [bff] Anubis client _has_script function returning "properly"

Yeah, if possible use function return values over string return.
String returns from functions is a thing but something that should be
avoided if possible, IMHO.

* [bff] Anubis client outputs DEBUG or VERBOSE over stderr

It is relatively common practice to bastardize stderr for general
output when writing scripts, especially when the string value return
is used from functions.  So stdout is "reserved" for function return
strings and stderr for everything else.

 - shrug - It's a BASH thing... :-/

let's move on ;-)

* [bff] _has_script - keep it functionally self contained...

Good catch Stas... Yes This function was not using what was passed
into it. It should.

Danke.