Bff telemetry (#623)
* [bff] Using runscript for launching application...

I want tigher, more fine-grained control over how this service is
launched. The command line will get more complex.  Complex enough that
I don't want to manage it via docker CMD array arguments.  So instead
will have Docker call the runscript that will pass through the
environment and launch the service.

This also makes it so that if you want to run the code yourself just
on your local machine you can run the launch script and you're good -
of course assuming you have the necessary infrastructure in place (or
stubbed).

* Set up RUN_SCRIPT to launch project & Including JMX in the launch.

I need to set values for memory for Java.  This level of command line
flexibility is necessary.  As this was being done I set up JMX to
export metrics.

Admittedly the JMX is a bit "extra credit" but I wanted to get the JVM
command line outta my face.