Doing some arg parsing gymnatics
It was bugging me that --force and --debug didn't work anywhere in the
command line.  There was also a need to a certain set of args ahead of
another set of args.  This created three groups of args... so to get
them all behaving appropriately I have to make multiple passes over
the args.  Here we have 3 passes over the args.
I cannot shift off of the main $@ array because I can't revisit them
once I have shifted off of them.  So I have to put $@ into an array
such that i can reset the index to traverse them again each time.
It looks a bit scary but not really.