Using Schema Validation for Descriptor
Using the luposlip/json-schema implementation.

https://github.com/luposlip/json-schema

Pros:
I like it because it explicitly supports schema v7
And it can take the JSON string or EDN (map) datastructure to
validate.  It also returns the input, so it simply passes the data
through if all good.  This means that I never actually have to turn
the TOML into JSON - I can keep it as map.  This saves a
transformation step and perturbs the code as little as possible.

Cons:
The only way to find out that something is broken is through an
exception that is not that informative.  Exceptions are expensive and
thus could have an adverse effect on the performance if bombarded with
bad descriptors.

Alternatives:
https://github.com/metosin/scjsv - It also can take the EDN (map) in
put of the descriptor data, but it does not have a pass through
function so that there wouldl be an additional step to adjudicate the
information. However the return output is very nice and informative.

https://github.com/niquola/json-schema.clj - This is also a nica
validator. It has clean cut boolean return for "valid?".  And it
returns friendlier output version when you call "validate".

We may want to switch over to the more informative versions if we want
to push #custmerobsession.  For this commit at least I'll take the
more unobtrusive path.